Restructure files for ReportPage components
