+++ frontend/src/components/reportForm/sections/Info.jsx
