+++ frontend/src/components/reportForm/sections/Photo.jsx
